+++ src/FluxFlow.Components.FileSystem/FileSystemComponentTypes.cs
@@ -4,5 +4,6 @@ namespace FluxFlow.Components.FileSystem;
 
 public static class FileSystemComponentTypes
 {
+    public static readonly NodeType FileRead = new("file.read");
     public static readonly NodeType FileWrite = new("file.write");
 }